+++ scribdl/__init__.py
@@ -2,8 +2,9 @@ from .version import __version__
 
 from .downloader import Downloader
 
-from .document import ScribdTextualDocument
-from .document import ScribdImageDocument
-from .book import ScribdBook
+from .content.document import ScribdTextualDocument
+from .content.document import ScribdImageDocument
+from .content.book import ScribdBook
+from .content.audiobook import ScribdAudioBook
 
 from .pdf_converter import ConvertToPDF
